style(internal): funcs to customize binary name/desc
**Internal Use Only**

To integrate the Soluble CLI into the Lacework CLI we are planning to add it as a
component (an extension) so that customers can discover and use it out-of-the-box.

These functions will allow us to inject the name of the component at runtime so that
all the help messages reflect the correct component name. If the environment variable
does not exist, this function defaults to the original name of the Soluble CLI.

Signed-off-by: Salim Afiune Maya <afiune@lacework.net>